requirements.txt
aiogram==3.7.0
aiohttp
python-dotenv
pillow
colorthief